Simplify IdleEvent.
- Return False when a ping has been sent, so current instance isn't trashed
  and replaced by ping's internal expectPacket.
- Make current idleEvent instance handle ping's timeout, to avoid creating
  a new event for each ping sent in relation to an existing event.
- Don't call expectPacket from ping when given a msg_id, meaning it is
  related to an existing event.
- Add a minimal delay of 5 seconds between 2 pings.
- Update event tests

git-svn-id: https://svn.erp5.org/repos/neo/trunk@1713 71dcc9de-d417-0410-9af5-da40c76e7ee4

+++ neo/connection.py
@@ -290,13 +290,17 @@ class Connection(BaseConnection):
                 return
             self.read_buf = self.read_buf[len(packet):]
 
+            packet_type = packet.getType()
             # Remove idle events, if appropriate packets were received.
             for msg_id in (None, packet.getId()):
                 event = self.event_dict.pop(msg_id, None)
                 if event is not None:
-                    self.em.removeIdleEvent(event)
+                    if packet_type == Packets.Pong:
+                        self.em.refreshIdleEvent(event)
+                        self.event_dict[msg_id] = event
+                    else:
+                        self.em.removeIdleEvent(event)
 
-            packet_type = packet.getType()
             if packet_type == Packets.Ping:
                 # Send a pong notification
                 self.answer(Packets.Pong(), packet.getId())
@@ -474,12 +478,13 @@ class Connection(BaseConnection):
         assert packet.isResponse(), packet
         self._addPacket(packet)
 
-    def ping(self, timeout=5):
+    def ping(self, timeout=5, msg_id=None):
         """ Send a ping and expect to receive a pong notification """
         packet = Packets.Ping()
-        msg_id = self._getNextId()
+        if msg_id is None:
+            msg_id = self._getNextId()
+            self.expectMessage(msg_id, timeout, 0)
         packet.setId(msg_id)
-        self.expectMessage(msg_id, timeout, 0)
         self._addPacket(packet)
